refactor all the duplicate tests
